+++ 1.21.5/common/src/main/java/net/xstopho/resource_cracker/item/ScytheItem.java
@@ -14,7 +14,7 @@ import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.xstopho.resource_cracker.config.ToolConfig;
 import net.xstopho.resource_cracker.item.components.TooltipContainer;
-import net.xstopho.resource_cracker.item.materials.ScytheToolMaterial;
+import net.xstopho.resource_cracker.item.materials.ScytheMaterial;
 import net.xstopho.resource_cracker.registries.DataComponentRegistry;
 
 import java.util.List;
@@ -23,7 +23,7 @@ import java.util.function.Supplier;
 public class ScytheItem extends Item {
     private static final Supplier<Integer> radius = () -> ToolConfig.scytheHarvestRadius;
 
-    public ScytheItem(ScytheToolMaterial material, float attackDamage, float attackSpeed, Properties properties) {
+    public ScytheItem(ScytheMaterial material, float attackDamage, float attackSpeed, Properties properties) {
         super(material.applyScytheProperties(properties, attackDamage, attackSpeed)
                 .component(DataComponentRegistry.TOOLTIP_CONTAINER.get(), new TooltipContainer(List.of(
                         Component.translatable("item.scythe.tooltip").withStyle(ChatFormatting.GOLD),